Use 'txnedit' task instead of 'simpleEdit' for updating transactions (#14)
* Use 'txnedit' task instead of 'simpleEdit' for updating transactions

The 'txnedit' task is able to update all properties of a transaction
including notes, tags, etc. It is also compatible with all fields
which can be submitted with the 'simpleEdit' task.

* Add integration test for editing a transaction

Note that you will need a integration-config.json file with valid
credentials and cookie information as these tests run against mint.com
itself.

* Increase timeout for integration test

* Only run integration tests if config is specified

Previous method using process.exit() would cause all test execution to
stop. We only want this test to not run, not for everything to halt in
the middle of test execution.

* Add tests demonstrating behavior of empty and undefined fields

These two new tests demonstrate that if you do not specify fields passed
to editTransaction() then the fields remain untouched.

- Large refactoring of integration tests

* Add semicolon to end of statement

* Fix and unignore test verifying undefined fields updated

The date field is required and must be passed to editTransaction()

* Remove unnecessary return statements to assertion methods

* Assert that note field has not changed as expected

We expect the note field to be the same as the original created
transaction after doing an update. This assertion is necessary because
we don't want mint clearing fields that we aren't passing and it's not
completely apparent what mints expected behavior is.

* Add semicolons to end of all statements

* Remove unneeded space between function and parameter list opening paren

* Indent integration-test with 4 instead of 2 spaces

* Add trailing comma to last element of data structures